Fixed adding item with event
Fixed the issue when adding items to the inventory using events. We were
firing the same event after adding the item to the inventory creating a
loop. Created a new event to be fired after the item is added to the
inventory and updated the UIInventory to listen for this event.